move linkedtrack to trackmanager
