+++ AudioCuesheetEditor/Services/AudioCuesheet/ITrackManager.cs
@@ -53,5 +53,19 @@ namespace AudioCuesheetEditor.Services.AudioCuesheet
         /// <param name="track"></param>
         /// <returns></returns>
         Track Clone(ITrack track);
+        /// <summary>
+        /// Get the previous linked track of a track object
+        /// </summary>
+        /// <param name="cuesheet">Cuesheet to get the previous linked track from</param>
+        /// <param name="track">Track object to get the previous link to</param>
+        /// <returns>Previous linked track or null (if not linked)</returns>
+        Track? GetPreviousLinkedTrack(Track track);
+        /// <summary>
+        /// Get next linked track that is linked to the parameter track
+        /// </summary>
+        /// <param name="cuesheet"></param>
+        /// <param name="track"></param>
+        /// <returns></returns>
+        Track? GetNextLinkedTrack(Track track);
     }
 }
